Updated install_cli and install_workflow scripts to better handle different shells with the help of Claude.
This was to address that the .env SAND_API_KEY wasn't getting properly exported to an env var since the command only worked within the context of the script running.

The `read -p` command was replaced with an echo + read command.

READMEs for docker_compose and workflow updated to include `source` when running the scripts, to ensure that the env var export works.